Create bar graph, ensure dictionary is altered
I wasn't sure if the ACTUAL dictionary was being altered when
the program stated it was being altered as the file is VS Code
stayed the same. That's because the file is pulled up once and
remains as it was it was pulled when it was opened. added
checks to shw that yes, it is being altered.

I forgot I needed to add a graph to anchan_popsort.py, so I did.
Learned a lot of new things.