Beginning work on matrices, not sure what changed
Apparently there are changes I made a while back, days ago.
Not sure what they are, but the file in question looks good and,
rather than risk mucking it up, I'm just gonna commit. I probably would
not do this in a more complicated project, but this is not that.